[refactor]: db 및 라우터 분리 - /chat/room/all - /chat/room/create
